+++ chirho/dynamical/handlers/event_loop.py
@@ -83,15 +83,3 @@ class InterruptionEventLoop(Generic[T], pyro.poutine.messenger.Messenger):
 
         msg["value"] = state
         msg["done"] = True
-
-    @staticmethod
-    def _pyro_simulate_to_interruption(msg) -> None:
-        from chirho.dynamical.internals.backends.torchdiffeq import (
-            torchdiffeq_simulate_to_interruption,
-        )
-
-        interruptions, dynamics, initial_state, start_time, end_time = msg["args"]
-        msg["value"] = torchdiffeq_simulate_to_interruption(
-            interruptions, dynamics, initial_state, start_time, end_time
-        )
-        msg["done"] = True
