+++ chirho/dynamical/handlers/solver.py
@@ -14,3 +14,15 @@ class TorchDiffEq(Solver):
             "options": options,
         }
         super().__init__()
+
+    @staticmethod
+    def _pyro_simulate_to_interruption(msg) -> None:
+        from chirho.dynamical.internals.backends.torchdiffeq import (
+            torchdiffeq_simulate_to_interruption,
+        )
+
+        interruptions, dynamics, initial_state, start_time, end_time = msg["args"]
+        msg["value"] = torchdiffeq_simulate_to_interruption(
+            interruptions, dynamics, initial_state, start_time, end_time
+        )
+        msg["done"] = True
